#127, 128, fixed login password not hidden and issues with leaving the game being called two times. Added Bug fixing task for project. Also added heartbeats for websocket to check if user closed browser

+++ src/main/java/ch/uzh/ifi/hase/soprafs26/service/GameFlowService.java
@@ -353,6 +353,11 @@ public class GameFlowService {
     public void leaveGame(String gameCode, String username) {
         Game game = getGameByCode(gameCode);
 
+        //guard against calling leaveGame twice
+        if (!game.getPlayers().containsKey(username)){
+            return;
+        }
+
         // Last player: delete the game regardless of status
         if (game.getPlayers().size() <= 1) {
             gameRepository.delete(game);
